refactor: migrate to Vite + ES modules
- Split monolithic web/script.js (2600 lines) into 10 ES modules in web/src/
- Replace CDN dependencies with npm packages (marked, mermaid, bootstrap, etc.)
- Add Vite as build tool with optimized chunking for mermaid diagrams
- Update desktop-app/prepare.js to use dist/ output from Vite
- Move desktop-specific code to desktop-app/src/desktop-main.js
- Add CLAUDE.md and AGENTS.md with updated architecture docs
- Add claude-skills package for development workflows

MathJax remains on CDN (config-based loading is simpler than bundling).
