+++ desktop-app/src/desktop-main.js
